some changes

+++ main.py
@@ -7,6 +7,7 @@ import ast
 import select
 import hashlib
 import shutil
+import random
 
 def get_ip_address():
     s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
@@ -16,6 +17,7 @@ def get_ip_address():
 HOST = get_ip_address()
 #FOR HAMACHI
 #HOST = "25.82.175.84"
+print(HOST)
 
 PORT = 12345
 
@@ -27,6 +29,7 @@ payload = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "DISCOVER", "PAYLOAD":""})
 response = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "RESPOND", "PAYLOAD":""})
 message = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "MESSAGE", "PAYLOAD":""})
 filePackage = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "FILE", "PAYLOAD":"", "serial":""})
+chunkPackage = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "CHUNK", "PAYLOAD":"", "serial": ""})
 AckPackage = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "ACK", "PAYLOAD":"", "serial":"", "rwnd": ""})
 
 goodbyeMessage = dict({"NAME":MYNAME, "MY_IP": HOST, "TYPE": "GOODBYE", "PAYLOAD": ""})
@@ -35,15 +38,18 @@ responseBytes = json.dumps((response)).encode('utf-8')
 goodbyeBytes = json.dumps((goodbyeMessage)).encode('utf-8')
 message_to_send=""
 acks_received = []
-
+chunk_amounts = {}
+chunk_amounts_user = {}
+is_transfer_done = {}
+list_of_chunks_transfered = {}
 
 #DISCOVER
 def discover():
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     sock.bind(('', 0))
     sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-    #sock.sendto(payloadBytes, ('<broadcast>', PORT))
-    sock.sendto(payloadBytes, ('25.255.255.255', PORT))
+    sock.sendto(payloadBytes, ('<broadcast>', PORT))
+    #sock.sendto(payloadBytes, ('25.255.255.255', PORT))
 
     print("Discover Broadcast sent")
 
@@ -110,13 +116,15 @@ def read_in_chunks(file_object, chunk_size):
 def create_temp(filename):
     if os.path.isfile('{}.txt'.format(str(filename))):
         with open('{}.txt'.format(str(filename)), "rb") as f:
-            shutil.rmtree('{}/{}_temp'.format(get_cwd(), str(filename)))
+            if (os.path.exists('{}_temp'.format(incomingData["FILENAME"]))):
+                shutil.rmtree('{}/{}_temp'.format(get_cwd(), str(filename)))
             os.mkdir('{}_temp'.format(str(filename)))
             os.chdir('{}/{}_temp'.format(get_cwd(), str(filename)))
             index = 0
             for chunk in read_in_chunks(f):
                 hash_ = get_hash(chunk)
                 chunk_file = open('{}.txt'.format(index), 'w+')
+                chunk = str(chunk, 'utf-8')
                 chunk_file.write(chunk)
                 chunk_file.close()
                 index = index + 1
@@ -127,15 +135,23 @@ def create_temp(filename):
         return False
 
 
-def request_chunk(filename, serial):
+def request_chunk_info(filename):
     sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
     sock.bind(('', 0))
     sock.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
     #sock.sendto(payloadBytes, ('<broadcast>', PORT))
+    packet = generate_packet("CHUNK", "", -1, "", filename)
+    chunk_amounts_user[filename] = []
+    packetBytes = json.dumps(packet).encode('utf-8')
+    sock.sendto(packetBytes, ('<broadcast>', PORT))
+
+def request_chunk(filename, destination_ip, serial):
+    sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    sock.bind(('', 0))
+    #sock.sendto(payloadBytes, ('<broadcast>', PORT))
     packet = generate_packet("FILEREQ", "", serial, "", filename)
     packetBytes = json.dumps((packet)).encode('utf-8')
-    sock.sendto(packetBytes, ('25.255.255.255', PORT))
-
+    sock.sendto(packetBytes, (destination_ip, PORT))
 
 def request_file():
     print("If you want to abort file request, type in ABORT")
@@ -143,19 +159,30 @@ def request_file():
     if filename == "ABORT":
         pass
     else:
-        completed_chunks = []
-        hash_values = {} # {index: hash} 
-        # get has values somehow
-        curr_hash = 0
-        #while (not(hash_values['fin'] == curr_hash)):
-        serial = 1
-        timer = time.time()
-        while not os.path.exists('{}_end.txt'.format(incomingData["FILENAME"])):
-            request_chunk(filename, serial)
-            serial = serial + 1
-            if time.time() - timer > 5:
-                print("Couldnt download file.")
-                break
+        request_chunk_info(filename)
+        time.sleep(1)
+        list_of_chunks_to_request = []
+        list_of_users_with_chunks = chunk_amounts_user[filename]
+        list_of_chunks_transfered[filename] = []
+
+        for user in list_of_users_with_chunks:
+            for chunk_id in user[1]:
+                list_of_chunks_to_request.append((user[0], chunk_id))
+
+        while not is_transfer_done[filename]:
+            
+            list_of_chunks_to_request = [i for pair in list_of_chunks_to_request (if pair[1] not in list_of_chunks_transfered)]
+
+            random_request = random.choice(list_of_chunks_to_request)
+            request_chunk(filename, random_request[0], random_request[1])
+
+#chunk_amounts_user = {"file1": [("userip1", [0, 1, 5, 6, 10]), ("userip2", [15 chunks])]   , "file2": ("userid2",5)}
+("userip1", 0)
+("userip1", 1)
+("userip1", 5)
+("userip1", 6)
+("userip1", 10)
+("userip2", 2)
 
 
 
@@ -169,7 +196,7 @@ def send_chunk(incomingData):
         for i in range(3):
             s.send(packetBytes)
             time.sleep(0.01)
-            if (incomingData["MY_ID"], incomingData["FILENAME"], incomingData["SERIAL"]) is not in acks_received:
+            if (incomingData["MY_ID"], incomingData["FILENAME"], incomingData["SERIAL"]) not in acks_received:
                 time.sleep(1)
             
 
@@ -195,6 +222,7 @@ def send_Message(name, messagePayload):
             s.close()
     except ConnectionRefusedError:
         print("unexpected offline client detected")
+
 
 
 def receiveTCP():
@@ -233,8 +261,17 @@ def sendACK(incomingData, leftBuffer):
     AckPackageBytes = json.dumps((AckPackage)).encode('utf-8')
     sock.sendto(payloadBytes, ('25.255.255.255', PORT))
 
-    print("Discover Broadcast sent")
-    pass
+    print("ack senttt!!!!")
+    
+
+def send_chunk_info(packet, receiver_ip):
+    sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    sock.bind(('', 0))
+    packetBytes = json.dumps((packet)).encode('utf-8')
+    sock.sendto(packet, (receiver_ip, PORT))
+    #sock.sendto(payloadBytes, ('25.255.255.255', PORT))
+
+    print("chunk info sent!!!")
 
 
 def putFileTogether(dataformat):
@@ -268,8 +305,32 @@ def receiveUDP():
                 dataDecoded = data.decode('utf-8')
                 incomingData = json.loads(dataDecoded)
                 if(incomingData["MY_IP"]== HOST):
-                    continue
+                    continue 
                 #print("receiving message")
+                if incomingData["TYPE"] == "CHUNK":
+                    if incomingData["SERIAL"] == -1:
+                        list_of_chunks = []
+                        if (os.path.exists('{}.txt'.format(incomingData["FILENAME"]))):
+                            create_temp()
+                            for filename in os.listdir('{}/{}_temp/'.format(get_cwd, incomingData["FILENAME"])):
+                                list_of_chunks.append(filename[:-3])
+                        elif (os.path.exists('{}/{}_temp/'.format(get_cwd, incomingData["FILENAME"]))):
+                            for filename in os.listdir('{}/{}_temp/'.format(get_cwd, incomingData["FILENAME"])):
+                                list_of_chunks.append(filename[:-3])
+                        else:
+                            pass
+                        sender_ip = incomingData["MY_IP"]
+                        respPackage = chunkPackage
+                        respPackage["PAYLOAD"] = list_of_chunks
+                        respPackage["SERIAL"] = len(list_of_chunks)                           
+                        send_chunk_info(respPackage, sender_ip)
+
+                    elif incomingData["SERIAL"] > 0:
+                        temp_list = chunk_amounts_user[incomingData["FILENAME"]]
+                        temp_list.append((incomingData["MY_IP"], incomingData["PAYLOAD"]))
+                        chunk_amounts_user[incomingData["FILENAME"]] = temp_list
+                        chunk_amounts["FILENAME"] = max(incomingData["SERIAL"], chunk_amounts["FILENAME"])
+
                 if incomingData["TYPE"] == "FILEREQ":
                     if (os.path.exists('{}_temp/{}.txt'.format(incomingData["FILENAME"], incomingData["SERIAL"]))):
                         send_chunk(incomingData)
@@ -287,6 +348,11 @@ def receiveUDP():
                             chunk_file = open('{}/{}.txt'.format(incomingData["FILENAME"], incomingData["SERIAL"]), 'wb+')
                             chunk_file.write(incomingData["PAYLOAD"])
                             chunk_file.close()
+                            temp_list = list_of_chunks_transfered["FILENAME"]
+                            temp_list.append(incomingData["SERIAL"])
+                            list_of_chunks_transfered[incomingData["FILENAME"]] = temp_list
+                            if len(os.listdir('{}/{}_temp/'.format(get_cwd, incomingData["FILENAME"]))) == chunk_amounts[incomingData["FILENAME"]]:
+                                is_transfer_done[incomingData["FILENAME"]] = True
                             #respond with ACK inclusing remaining buffer --> here hardcoded
                             sendACK(incomingData,1500)
                 if (incomingData["TYPE"]=="DISCOVER"):
@@ -351,7 +417,6 @@ while(True):
         print("no chat partner found yet")
         time.sleep(10)
         continue
-
     showPartners("SHOW")
     message_to_send =""
     partnerName = input("To send a Message type in the Name of a Chat partner")
@@ -359,7 +424,7 @@ while(True):
     #print("this is his NAme: "+partnerName)
     if partnerName == "FILE":
         request_file()
-    elif: partnerName == "HELP":
+    elif partnerName == "HELP":
         display_help()
     elif partnerName in discoveredUsers:
         print(("please type in your message (to go back to the menu type in EXIT)"))
